replace SplitList with Split
If my hunch is correct, this call to isSubPath currently does not work as intended.
SplitList splits by the OS ListSeperator, which is usually `:`. I'm thinking that this call was supposed to be for paths and not a list of separated paths considering it was called to parse through the paths in the vdf files in `getDataToMove`. I'm adding what I think was the intended functionality of this code, which is splitting it by the path separator.